requirements.txt
discord.py[voice] ~= 0.12.0
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
schedule

